x/blueprints/types/genesis_test.go
package types_test

import (
	"testing"

	"github.com/rafaelsousa/realestate/x/blueprints/types"
	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	for _, tc := range []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				HouseList: []types.House{
					{
						Id: 0,
					},
					{
						Id: 1,
					},
				},
				HouseCount: 2,
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated house",
			genState: &types.GenesisState{
				HouseList: []types.House{
					{
						Id: 0,
					},
					{
						Id: 0,
					},
				},
			},
			valid: false,
		},
		{
			desc: "invalid house count",
			genState: &types.GenesisState{
				HouseList: []types.House{
					{
						Id: 1,
					},
				},
				HouseCount: 0,
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	} {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
